Update to v4.0.0-beta.9
I finally feel like I'm getting somewhere with these updates!

Moved many class members to be private with get / set functions to get /
set the values. I left a few not like this because it didn't make sense
might change that later.

Changed show map to have an edge of 0.