Update to v4.0.0-rc.1
Finally it seems to be working! (Sorta)

Made all the class members that were private to public and this fixed
everything! I don't understand why I needed to do this; but I'm ready to
be done with v4.0.0

A few clean ups are in order as well as a few small bugs I need to iron
out but for the most part its working