Spilt Compile into Final & Debug
Spilt Complie.bat into a Complie-Final.bat which will compile the final
program and Complie-Debug.bat which will compile a debug version. The
only differences are in the options. Debug will have warnings show as
well as profiler and debuger. While the Final will be compiled for speed
and mininal file size.

Changes also reflected in Makefile, Object-Builder and Complie.sh